refactor : 구입 개수 기능을 money에서 lottomachine으로 이동
